fix CWInitialization behaviour
CWInitialization used to generate a permutation and use it as is. It
means that when the instance was printed with this permutation, both the
rows and columns were permuted.

In our case, we only want rows to be permuted to generate the CW
initialization. In order to do that, an Instance::permuteRows() method
was created which takes a Permutation as a parameter. When we call the
generateInitialization() from CWInitialization, the following steps are
followed:

1. Generate permutation (of the rows)
2. alter the instance with permuteRows() and the previously generated
permutation
3. return a base Permutation (0, 1, ..., n)