First/BestImprovement compute new permutation constraint
Only compute a new permutation if its score is superior to the current
permutation. This modification is possible because Neighbourhood::dela()
does not require the new permutation to evaluate it, it only needs the
current permutation to compute the score for the new one. Thus,
Instance::evaluate() do not require the new permutation neither.